Move command reader into standalone module
