+++ connector/pom.xml
@@ -37,25 +37,6 @@
             <scope>compile</scope>
         </dependency>
         <dependency>
-            <groupId>net.minecrell</groupId>
-            <artifactId>terminalconsoleappender</artifactId>
-            <version>1.0.0</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <version>1.18.4</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.fusesource.jansi</groupId>
-            <artifactId>jansi</artifactId>
-            <version>1.18</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
             <groupId>com.nukkitx.protocol</groupId>
             <artifactId>bedrock-v388</artifactId>
             <version>2.4.0</version>
@@ -110,6 +91,12 @@
                     <artifactId>mcauthlib</artifactId>
                 </exclusion>
             </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.4</version>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
     <build>
